Support files > 2GB in Win32 builds by default
Define _FILE_OFFSET_BITS to 32 to go back to the old (32-bit off_t) behavior.
There's still some sticky wickets here (credits are still 32-bit, etc.), but it's a start.

+++ src/sbbs3/bat_xfer.cpp
@@ -125,7 +125,7 @@ void sbbs_t::batchmenu()
 						bprintf(text[DownloadQueueLstFmt],i+1
 							,filename
 							,ultoac(f.cost, tmp)
-							,ultoac(f.size, str)
+							,ultoac((ulong)f.size, str)
 							,cur_cps
 								? sectostr((uint)(f.size/(ulong)cur_cps),tmp2)
 								: "??:??:??"
