+++ src/sbbs3/qwk.cpp
@@ -580,11 +580,11 @@ void sbbs_t::qwk_sec()
 
 			off_t l=flength(str);
 			bprintf(text[FiFilename], getfname(str));
-			bprintf(text[FiFileSize], ultoac(l,tmp)
+			bprintf(text[FiFileSize], ultoac((ulong)l,tmp)
 				, byte_estimate_to_str(l, tmp2, sizeof(tmp), /* units: */1024, /* precision: */1));
 
 			if(l>0L && cur_cps)
-				i=l/(ulong)cur_cps;
+				i=(uint)(l/(ulong)cur_cps);
 			else
 				i=0;
 			bprintf(text[FiTransferTime],sectostr(i,tmp));
